% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundsGamma.R
\name{boundsGamma}
\alias{boundsGamma}
\title{Upper limit for overall mutation rate}
\usage{
boundsGamma(alleles, afreq, range)
}
\arguments{
\item{alleles}{A character vector with allele labels.}

\item{afreq}{A numeric vector of allele frequencies.}

\item{range}{A positive number.}
}
\value{
Bounds on gamma
}
\description{
Upper limit for overall mutation rate
}
\examples{
alleles = 1:3; afreq = c(0.2, 0.3,  0.5);  range = 0.1
bounds = boundsGamma(alleles, afreq ,range)
                        
}
\author{
Thore Egeland.
}
