% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLRR.R
\name{compareLRR}
\alias{compareLRR}
\title{Ratio of LRs are compared for different tranformations}
\usage{
compareLRR(M, R, ped = nuclearPed(1), ids = c(1, 3), ln = FALSE)
}
\arguments{
\item{M}{List of mutations matrices, typically not reversed}

\item{R}{List of reversed mutations matrices}

\item{ped}{A \code{ped} object}

\item{ids}{A numeric with ID labels 
of one or more pedigree members.}

\item{ln}{character  Logical.}

\item{method}{Character specifying reversing method.}

\item{adjust}{character  Logical.}
}
\value{
Statistcs evaluation transformation
}
\description{
The ratio Z = LR(M,p)/LR(R,p) is calculated. for a SNP marker
}
\examples{
ped = nuclearPed(1)
p = c("1" = 0.2, "2" = 0.8)
M = mutationMatrix("equal", afreq = p, alleles = 1:length(p), rate = 0.01)
R = makeReversible(M, method = "PM", afreq = p)
M = list(M)
names(M) = "L1"
R = list(R)
res1 = compareLRR(M,R)

M =  R = list()
M[[1]] = mutationMatrix("equal", afreq = p, alleles = 1:length(p), rate = 0.01)
M[[2]] = mutationMatrix("equal", afreq = p, alleles = 1:length(p), rate = 0.05)
R[[1]] = makeReversible(M[[1]], method = "PM", afreq = p)
R[[2]] = makeReversible(M[[2]], method = "PM", afreq = p)
names(M) = names(R) = paste0("L", 1:2)
res2 = compareLRR(M, R)
res3 = compareLRR(M, R, ln = TRUE)
}
\seealso{
[makeReversible()].
}
\author{
Thore Egeland.
}
