% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distLRR.R
\name{distLRR}
\alias{distLRR}
\title{Find summary statistics of LRR}
\usage{
distLRR(
  adjust = TRUE,
  method = "PR",
  verbose = FALSE,
  nsim = 2,
  seed = NULL,
  log2 = FALSE,
  mutMod = "equal",
  rate = c(0.005, 0.005),
  rate2 = c(1e-06, 1e-06),
  range = c(0.1, 0.1),
  nAls = 4
)
}
\arguments{
\item{adjust}{Logical. Transform to original mutation rate}

\item{method}{Transformation. See `pedmut::makeReversible()`}

\item{verbose}{Logical.}

\item{nsim}{integer.}

\item{seed}{Integer}

\item{log2}{Logical. To get LR on log2}

\item{mutMod}{character.}

\item{rate}{Double. See `pedmut::muationModel()`.}

\item{rate2}{Double. See `pedmut::muationModel()`}

\item{range}{Double. See `pedmut::muationModel()`}

\item{nAls}{Integer. No of alleles.}
}
\value{
Summary of distribution of LRR, with more details if `verbose = T`
}
\description{
Finds distribution of LRR depending on allele frequencies.
}
\details{
The allele frequencies are simulated uniformly. The mutation
parameters are simulated uniformly between provided minimum and maximum.
The LRR distribution is then calculated exactlyusing
`oneMarkerDistribution()`. The simulations are conditional on PR not failing
(PR fails rarely).#'
}
\examples{
distLRR()

}
\author{
Thore Egeland
}
