% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactLRR.R
\name{exactLRR}
\alias{exactLRR}
\title{Ratio of LRs and likelihoods calculated exactly for one marker}
\usage{
exactLRR(
  M,
  R = NULL,
  afreq = NULL,
  ped = nuclearPed(1),
  ids = c(1, 3),
  method = "PR",
  adjust = FALSE,
  ln = FALSE
)
}
\arguments{
\item{M}{Mutation matrix}

\item{R}{Mutation matrix with same dimension as M or NULL}

\item{afreq}{A vector with allele frequencies, 
of the same length as the size of mutmat}

\item{ped}{A \code{ped} object}

\item{ids}{A numeric with ID labels 
of one or more pedigree members.}

\item{method}{Character specifying reversing method.}

\item{adjust}{Logical.}

\item{ln}{Logical.}
}
\value{
Expected ratio and likelihoods
}
\description{
The ratio Z = LR(M,p)/LR(R,p) is calculated exactly.
}
\details{
If \code{R == NULL}, \code{R} is first made reversible with
\code{PR} option, preserving the expected mutation rate.
}
\examples{
ped = nuclearPed(1)
p = c("1" = 0.2, "2" = 0.8)
M = mutationMatrix("equal", afreq = p, alleles = 1:length(p), rate = 0.01)
R = findReversible(M, method = "PR", afreq = p)
exactLRR(M, R, ln = FALSE)

\dontrun{
p = forrel::NorwegianFrequencies[[1]]
M = mutationMatrix("onestep",
    alleles = names(p),
    rate = 0.001,
    afreq = p)
attr(M, "rate") = expectedMutationRate(M, p)
res = exactLRR(M, R = NULL, ln = T)
R = findReversible(M, afreq = p, method = "MH")
R2 = adjustReversible(M, R, afreq = p)
res2 = exactLRR(M, R = R, p)
}
}
\seealso{
[findReversible()].
}
\author{
Thore Egeland.
}
