% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedMutationRate.R
\name{expectedMutationRate}
\alias{expectedMutationRate}
\title{Expected mutation rate of a mutation matrix}
\usage{
expectedMutationRate(mutmat, afreq = NULL, check = TRUE)
}
\arguments{
\item{mutmat}{A mutation matrix.}

\item{afreq}{A vector with allele frequencies 
of the same length as the size of mutmat.}

\item{check}{Logical.}
}
\value{
Expected mutation rate.
}
\description{
Expected mutation rate of a mutation matrix
}
\examples{
library(pedmut)
n = 4
p= 1:n/sum(1:n)
names(p) = 1:n
mutmat = mutationMatrix("onestep", rate = 0.02, afreq = p, alleles = 1:n)
expectedMutationRate(mutmat, afreq = p)

}
\author{
Thore Egeland
}
