% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fratio.R
\name{fratio}
\alias{fratio}
\title{Calculates fratio}
\usage{
fratio(A, B)
}
\arguments{
\item{A}{Mutation matrix}

\item{B}{Mutation matrix}
}
\value{
Double
}
\description{
Calculates fratio, distance between matrices
}
\examples{
library(pedtools)
library(pedmut)
n = 2
p = c(0.2, 0.8)
names(p) = paste(1:n)
mutmat = mutationMatrix("equal", rate = 0.02, alleles = 1:2, afreq = p)
res = snpExample(mutmat, afreq = p, adjust = FALSE)
fratio(mutmat, res[[1]])

}
\author{
Thore Egeland <Thore.Egeland@nmbu.no>
}
