% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd1.parental.R
\name{ibd1.parental}
\alias{ibd1.parental}
\title{Estimates for a pair of non-inbred individuals the probabilities of parental
origin when IBD is 1}
\usage{
ibd1.parental(x, id.pair, Nsim, cM = 10000, verbose = F)
}
\arguments{
\item{x}{A `ped` object.}

\item{id.pair}{Integer vector of length 2 giving the pair of individuals.}

\item{Nsim}{Integer. The number of simulations.}

\item{cM}{Double. The length of the `uniformMap`.}

\item{verbose}{Logical.}
}
\value{
The averaged values.
}
\description{
Assume IBD is 1 for a pair of non-inbred individuals. The function estimates
the probabilities of the four possible combinations of parental origin: (i)
paternal-paternal, (ii) paternal-maternal, (iii) maternal-paternal and (iv)
maternal-maternal.
}
\examples{

library(ibdsim2)
x = doubleFirstCousins()
ids = leaves(x)
ibd1.parental(x, ids, 10)
}
\author{
Magnus Dehli Vigeland and Thore Egeland.
}
