% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isWellBehaved.R
\name{isWellBehaved}
\alias{isWellBehaved}
\title{Checks if mutation model is well behaved}
\usage{
isWellBehaved(Q, afreq)
}
\arguments{
\item{Q}{a mutation model}

\item{afreq}{A numeric vector of allele frequencies.}
}
\value{
A logical
}
\description{
Checks if q_ij <= p_j
}
\examples{
\dontrun{
# Not well behaved:
p = c(0.4, 0.6)
Q = stepwiseReversible(alleles = 1:2, afreq = p, rate = 0.6, range = 0.1)
Q = Q$M
isWellBehaved(Q, afreq = p)
}
}
\author{
Thore Egeland.
}
