% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRmutBounds.R
\name{LRmutBounds}
\alias{LRmutBounds}
\title{Pairwise LR bounds for proportional mutation models}
\usage{
LRmutBounds(p, kappa, K = 0, share = TRUE)
}
\arguments{
\item{p}{Vector of real numbers. Allele frequency vector.}

\item{kappa}{Vector of real numbers describing relationship.  IBD parameters
for 0,1,2 IBD alleles.}

\item{K}{Real. Proportionality factor in proportional model}

\item{share}{Logical. If TRUE, both are homozygous for the same allele, 
otherwise no akllele sharing.}
}
\value{
LR and the lower bound,  assuming no allele sharing, and 
the upper bound assuming the individuals to be homozygous for the rarest allele.
}
\description{
Reversibility is assumed for the mutation model and the LR is
calculated for a pair of non-inbred individuals comparing a kappa to unrelated.
}
\examples{

# Parent offspring relationship LR
p = c("1" = 0.2, "2" = 0.8)
M = mutationMatrix("proportional", afreq = p, rate = 0.003)
gamma = mut2::expectedMutationRate(M, p)
K = gamma/(1- sum(p^2))
kappa = c(0, 1, 0)
LRmutBounds(p, kappa, K, share = FALSE)
LRmutBounds(p, kappa, K, share = TRUE)
}
\references{
Egeland, Pinto and Amorim, FSI:Genetics (2017),
  \doi{http://dx.doi.org/10.1016/j.fsigen.2017.04.018}.
}
\author{
Thore Egeland <Thore.Egeland@nmbu.no>
}
