% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRR.R
\name{LRR}
\alias{LRR}
\title{Expected ratio of LRs}
\usage{
LRR(Q, M, afreq, ped, ids)
}
\arguments{
\item{Q}{Mutation matrix}

\item{M}{Mutation matrix with same dimension as Q}

\item{afreq}{A vector with allele frequencies, 
of the same length as the size of mutmat}

\item{ped}{A \code{ped} object}

\item{ids}{A numeric with ID labels 
of one or more pedigree members.}
}
\value{
Expected ratio and ratios
}
\description{
The ratio Z = LR(Q,p)/LR(M,p) and its 
expected value is calculated.
}
\details{
If \code{M == NULL}, \code{M} is first made reversible with
\code{PM} option, preserving the expected mutation rate.
}
\examples{
ids = c("id1", "id2")
ped = nuclearPed(father = ids[1], children = ids[2])
p = c("1" = 0.2, "2" = 0.8)
Q = mutationMatrix("equal", afreq = p, alleles = 1:length(p), rate = 0.01)
Q = mutationMatrix("onestep", afreq = p, alleles = 1:length(p), rate = 0.01)
M = findReversible(Q, method = "PM", afreq = p)
LRR(Q, M, p, ped, ids)
}
\seealso{
[findReversible()].
}
\author{
Thore Egeland.
}
