% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut2-package.R
\docType{package}
\name{mut2-package}
\alias{mut2}
\alias{mut2-package}
\title{mut2: Mutation stuff}
\description{
Making mutation models reversible and testing.
}
\author{
\strong{Maintainer}: Thore Egeland \email{thore.egeland@nmbu.no} (\href{https://orcid.org/0000-0002-3465-8885}{ORCID})

}
\keyword{internal}
