% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasonableMutationMatrix.R
\name{reasonableMutationMatrix}
\alias{reasonableMutationMatrix}
\title{Checks if pj < mij}
\usage{
reasonableMutationMatrix(mutmat, p, verbose = FALSE)
}
\arguments{
\item{mutmat}{A mutation matrix.}

\item{verbose}{Logical.}

\item{afreq}{A vector with allele frequencies 
of the same length as the size of mutmat.}
}
\value{
Number of cases where pj < mij
}
\description{
Entries where  pj < mij are unreasonable and are detected
}
\examples{
n = 4
p = 1:n/sum(1:n)
names(p) = 1:n
mutmat = mutationMatrix("equal", afreq = p, rate = 0.5)
reasonableMutationMatrix(mutmat, p, verbose = F)

}
\author{
Thore Egeland
}
