% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLR.R
\name{simLR}
\alias{simLR}
\title{Simulate LRRR}
\usage{
simLR(M, R, ped1 = nuclearPed(1), ids = c(1, 3), nsim = 2, seed = NULL)
}
\arguments{
\item{M}{List of mutation matrices, one for each marker.}

\item{R}{List of reversed mutation matrices, one for each marker.}

\item{ped1}{A \code{ped} object.}

\item{ids}{A numeric with ID labels of one or more pedigree members.}

\item{nsim}{Integer.}

\item{seed}{Integer.}
}
\value{
LRR
}
\description{
The ratio Z = LR(M,p)/LR(R,p) is simulated where 
M is a list of mutation matrices and R a reversed version
}
\details{
If \code{R == NULL}, mutation matrices are reversed.
}
\examples{

p = c("1" = 0.1, "2" = 0.9)
M = list(mutationMatrix("equal", rate = 0.001, afreq = p),
         mutationMatrix("equal", rate = 0.001, afreq = p))
mN = c("L1", "L2")
markerNames = mN; method = "PM"; ped1 = nuclearPed(1); ids = c(1,3); nsim = 2; seed = 17
res = simLRR(M, markerNames = mN, nsim = 100, seed = 177)
simLRR(M, markerNames = mN, nsim = 10, seed = 177)
res = simLRR(M[[1]], markerNames = mN[1], nsim = 10, seed = 177)
\dontrun{
M = lapply(NorwegianFrequencies, function(x){
  mat = mutationMatrix("stepwise",
  alleles = names(x),
  rate = 0.001, 
  rate2 = 1e-05,
  range = 0.1,
  afreq = x)
  mat
})
simLRR(M, markerNames = mN, nsim = 10, seed = 177)
}
}
\seealso{
[makeReversible()].
}
\author{
Thore Egeland.
}
