% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLRR.R
\name{simLRR}
\alias{simLRR}
\title{Simulate LRR}
\usage{
simLRR(
  M,
  R = NULL,
  markerNames,
  method = "PR",
  ped1 = nuclearPed(1),
  ids = c(1, 3),
  nsim = 2,
  seed = NULL
)
}
\arguments{
\item{M}{List of mutation matrices, one for each marker.}

\item{R}{List of reversed mutation matrices, one for each marker.}

\item{markerNames}{Character vector, names of markers}

\item{method}{Character, reversing method.}

\item{ped1}{A \code{ped} object.}

\item{ids}{A numeric with ID labels of one or more pedigree members.}

\item{nsim}{Integer.}

\item{seed}{Integer.}
}
\value{
LRR
}
\description{
The ratio Z = LR(M,p)/LR(R,p) is simulated where
M is a list of mutation matrices and R a reversed version.
The alternative hypothesis is that all individuals are unrelated and
so it does not matter if matrix M or R is used.
}
\details{
If \code{R == NULL}, mutation matrices are reversed.
}
\examples{
library(pedmut)
library(pedtools)
p = c("1" = 0.1, "2" = 0.9)
M = list(mutationMatrix("equal", rate = 0.001, afreq = p),
         mutationMatrix("equal", rate = 0.001, afreq = p))
mN = c("L1", "L2")
markerNames = mN; method = "PR"; ped1 = nuclearPed(1); ids = c(1,3); nsim = 2; seed = 17
res = simLRR(M, markerNames = mN, nsim = 10, seed = 177)
res = simLRR(M[[1]], markerNames = mN[1], nsim = 2, seed = 177)
\dontrun{
M = lapply(NorwegianFrequencies, function(x){
  mat = mutationMatrix("stequal",
  alleles = names(x),
  rate = 0.001,
  rate2 = 0,
  range = 0,
  afreq = x)
  mat
})
mN = names(NorwegianFrequencies)
simLRR(M, markerNames = mN, nsim = 10, seed = 177)
}
}
\seealso{
[findReversible()].
}
\author{
Thore Egeland.
}
