% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpExample.R
\name{snpExample}
\alias{snpExample}
\title{SNP example}
\usage{
snpExample(mutmat, afreq = NULL, check = TRUE, adjust = TRUE)
}
\arguments{
\item{mutmat}{A mutation matrix.}

\item{afreq}{A vector with allele frequencies}

\item{check}{Logical}

\item{adjust}{Logical}
}
\value{
Balanced mutation matrix. The expected mutation rate
of the balanced matrix is returned as `rate`.
}
\description{
The three reversible transformations and exact
}
\examples{
library(pedmut)
n = 2
p = c("1" = 0.2, "2" = 0.8)#' 
m = rbind(c(0.997, 0.003), c(0.003, 0.997))
mutmat = mutationMatrix("custom", matrix  = m, afreq = p, alleles = 1:2)
res = snpExample(mutmat, afreq = p, adjust = FALSE)

}
\author{
Thore Egeland
}
