% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabExact.R
\name{tabExact}
\alias{tabExact}
\title{Produce table comparing transformations based on exact calculations}
\usage{
tabExact(
  method = "PR",
  db = NorwegianFrequencies,
  markers = c("TH01", "D5S818", "D13S317", "D16S539", "TPOX", "D10S1248", "D22S1045",
    "D3S1744", "D2S1360", "D6S474", "D4S2366", "D5S2500", "D10S2325"),
  adjust = T,
  rate = 0.003,
  rate2 = 0.001,
  range = 0.5
)
}
\arguments{
\item{method}{Character. 'MH','PR' or 'BA'}

\item{db}{Database}

\item{markers}{character vector}

\item{adjust}{logical}

\item{rate}{Double}

\item{rate2}{Double}

\item{range}{Double}
}
\value{
Table (tab) and complete output (all)
}
\description{
Produce table
}
\details{
None
}
\examples{
library(pedsuite)
res = tabExact(method = 'PR')$tab

}
\author{
Thore Egeland
}
