% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabfRatio.R
\name{tabfRatio}
\alias{tabfRatio}
\title{Produce table comparing transformations based on exact calculations}
\usage{
tabfRatio(
  db = NULL,
  rate = 0.001,
  rate2 = 1e-05,
  range = 0.1,
  mutmodel = "stepwise",
  relabel = F,
  stationary = T,
  nr = 0
)
}
\arguments{
\item{db}{allele frequencies}

\item{rate}{Double}

\item{rate2}{Double}

\item{range}{Double}

\item{mutmodel}{Character}

\item{relabel}{Logical}

\item{stationary}{Logical}

\item{nr}{Integer 0 complete, 1,2,3 gives Table 1,2,3}
}
\value{
Table (tab) and complete output (all)
}
\description{
Produce table
}
\details{
If logical is TRUE, a stationary version is found and its fratio
to original matrix and also a check is performed to if it is also reversible
}
\examples{
library(pedsuite)
library(pedmut)
tabfRatio(db = NorwegianFrequencies[1:2], rate = 0.0001, mutmodel = 'equal')
}
\author{
Thore Egeland
}
