% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabLRR.R
\name{tabLRR}
\alias{tabLRR}
\title{Produce table comparing transformations based on simulations}
\usage{
tabLRR(
  db = NULL,
  rate = 0.001,
  rate2 = 1e-05,
  range = 0.1,
  nsim = 2,
  seed = 1729,
  maks = FALSE
)
}
\arguments{
\item{db}{allele frequencies}

\item{rate}{Double}

\item{rate2}{Double}

\item{range}{Double}

\item{nsim}{Integer}

\item{seed}{Integer}

\item{maks}{Logical}
}
\value{
LRR table
}
\description{
Produce table
}
\details{
LRR table
}
\examples{
\dontrun{
library(forrel)
library(KLINK)
db = getFreqDatabase(KLINK::halfsib[[1]])
tabLRR(db = db)
}
}
\author{
Thore Egeland
}
