% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustReversible.R
\name{adjustReversible}
\alias{adjustReversible}
\title{Adjusts Metropolis - Hasting balanced mutation matrix}
\usage{
adjustReversible(
  mutmat,
  balancedMutmat,
  method = "MH",
  afreq = NULL,
  check = TRUE
)
}
\arguments{
\item{mutmat}{Original mutation matrix.}

\item{balancedMutmat}{Balanced, mutation matrix. If, NULL, balanced.}

\item{method}{Character.  'MH','PR' or 'BA'.}

\item{afreq}{A vector with allele frequencies. 
of the same length as the size of mutmat.}

\item{check}{Logical.}
}
\value{
Adjusted mutation matrix
}
\description{
The Metropolis - Hastings conversion typically gives a mutation matrix with too small
small mutation rate. The previously balanced mutation matrix is adjusted
to have the same mutation rate as the original matrix.
}
\details{
If \code{balancedMutmat == NULL}, \code{mutmat} is first balanced.
The adjusted balanced matrix is

\code{alpha * balancedMutmat + (1-alpha) * I}

where

\code{alpha} is the ratio of the (expected mutation) rates of the original matrix,
\code{mutmat} to the balanced version \code{balancedMutmat} and \code{I} is the identity matrix.
}
\examples{
library(pedmut)
afreq = c(0.1, 0.3, 0.4, 0.2)
names(afreq) = 1:4
mutmat = mutationMatrix("onestep", alleles = 1:4, rate = 0.02)
balancedMutmatMH = findReversible(mutmat, afreq = afreq, method = "MH")
adj = adjustReversible(mutmat, balancedMutmatMH, afreq = afreq,  check = TRUE)
attr(mutmat, "rate") - attr(adj, "rate")

# Does 'PR' give adjusted matrix directly?:
balancedMutmatPR = findReversible(mutmat, afreq = afreq, method = "PR")
}
\seealso{
[findReveversible()].
}
\author{
Thore Egeland.
}
