% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distLRR.R
\name{distLRR}
\alias{distLRR}
\title{Find summary statistics of LRR}
\usage{
distLRR(
  adjust = TRUE,
  method = "PR",
  verbose = FALSE,
  nsim = 2,
  seed = NULL,
  log2 = FALSE,
  mutMod = "equal",
  Rate = c(0.005, 0.005),
  Rate2 = c(1e-06, 1e-06),
  Range = c(0.1, 0.1),
  nAls = 4
)
}
\arguments{
\item{adjust}{Logical.}

\item{method}{Transformation.}

\item{verbose}{Logical.}

\item{nsim}{integer.}

\item{seed}{Integer}

\item{log2}{Logical.}

\item{mutMod}{character.}

\item{Rate}{double.}

\item{Rate2}{Double.}

\item{nAls}{Integer.}

\item{range2}{Double.}
}
\value{
Summary of distribution of LRR
}
\description{
Finds distribution of LRR depending on allele frequencies.
}
\details{
The allele frequencies are simulated. The LRR distribution is then
exactly calculated using oneMarkerDistribution. RMSD (Root Mean Square
Deviation) is the square root of
E(LRR - E(LRR))^2, very close to SD(LRR) in our applications.
}
\examples{
distLRR()

}
\author{
Thore Egeland
}
