% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwiseReversible.R
\name{stepwiseReversible}
\alias{stepwiseReversible}
\title{Reversible stepwise mutation model}
\usage{
stepwiseReversible(alleles, afreq, rate, range)
}
\arguments{
\item{alleles}{A character vector with allele labels.}

\item{afreq}{A numeric vector of allele frequencies.}

\item{rate}{A numeric mutation rate.}

\item{range}{A positive number.}
}
\value{
A reversible stepwise mutation model with expected mutation rate equal input rate.
}
\description{
Reversible stepwise mutation model
}
\details{
The approach of Dawid el al. (SJS, 2001) is implemented.
}
\examples{
stepwiseReversible(alleles = 1:3, afreq = c(0.001, 0.499,  0.5), rate = 0.001, range = 0.1)

}
\author{
Thore Egeland.
}
